requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
opencv-python-headless>=4.9.0.80
numpy>=1.26.0
Pillow>=10.3.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
pywin32>=306; sys_platform == "win32"
python-dotenv>=1.0.0